requirements.txt
Django                       4.0.5
Pillow                       9.1.1
pip                          22.0.4
PyMySQL                      1.0.2
django-ckeditor              6.4.2
cryptography                 37.0.4 